Misc: Make the default to not use compression and add support for GZIP
A previous commit changed the default compression to use LZF instead of
GZIP. This leads to very nice speed improvements for motion correction.
However, LZF-compressed files are not supported by Fiji's HDF5 plugins.

This commit therefore makes the default to not use compression (GZIP too
slow, LZF not enough support) but allows specifying GZIP or LZF if the
user wishes.