Meta: Switch to using `uv` for package management
This changes from using `hatch` exclusively for managing development
environments to using `uv`. The build backend remains `hatchling` but
most of the environment-related things are moved over to `uv`. This is
done automagically by defining a 'dev' entry in the `dependency-groups`
array which `uv` uses as the default group it creates an environment
for.

The only step that still requires a `hatch` command is when running
'scripts' (ready-made shell commands) that are invoked with `hatch run
<COMMAND>`. Since `uv` does not provide a mechanism by which to lay out
regularly-used shell commands, usage should remain to use `hatch`.
For that reason, the 'default' hatch environment has been given the path
to the `uv` venv and all the 'scripts' have been moved to the 'default'
environment, but are now prefixed by their group name.

So where before, one would type-check using `hatch run types:check`,
one now runs it as `hatch run types-check`.

Currently, no `uv`-specific configuration is needed in the
`pyproject.toml` (i.e., no need for a `[tool.uv]` array) which has the
positive side effect of making the project more open to other backends
(e.g., switching to `setuptools`).