Lint: Add linting using `ruff`
`ruff` is almost fully-compatible with `black` formatting, with a few
deviations (that I happen to agree with for the most part).

Linting was missing from the `pyproject.toml` and this commit adds
`ruff` instead of `black` mainly for its speed when used as a language
server but also because of its added functionality through `ruff check`.

The current `ruff` configuration is shamelesly taken from the `mypy`
`pyproject.toml` file and will most likely be adapted in the future.