Lint: Fix `G` rules and disable `G001`, `G002`, and `G004`
I'm not convinced we're losing much eagerly evaluating formatted strings
in logging calls. The alternative, using `extra` or *args, makes the
calls much more cumbersome and less readable (more lines and requires
the use of %-formatting for positional arguments which I find horrible).

If we notice that eagerly formatting strings for logging levels that are
never used becomes a performance bottleneck, we can revisit this.